Fixed pr using updated code from main including routing
